+++ tests/utils/application.py
@@ -1,11 +1,14 @@
+from typing import Optional
+
+from blacksheep.messages import Request, Response
 from blacksheep.server import Application
 
 
 class FakeApplication(Application):
     def __init__(self, *args, **kwargs):
         super().__init__(show_error_details=True, *args, **kwargs)
-        self.request = None
-        self.response = None
+        self.request: Optional[Request] = None
+        self.response: Optional[Response] = None
 
     def normalize_handlers(self):
         if self._service_provider is None:
@@ -18,8 +21,8 @@ class FakeApplication(Application):
         self.normalize_handlers()
 
     async def handle(self, request):
-        self.request = request
         response = await super().handle(request)
+        self.request = request
         self.response = response
         return response