+++ CHANGELOG.md
@@ -5,13 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.1] - 2021-10-31 :shield:
+## [1.2.1] - 2021-??-?? :shield:
 - Adds built-in support to `JWT` bearer authentication, and validation
   of `JWTs` issued by identity providers implementing **OpenID Connect (OIDC)**
   discovery `/.well-known/openid-configuration` (more in general, for JWTs
   signed using asymmetric encryption and verified using public RSA keys)
+- Adds built-in support for **OpenID Connect (OIDC)** **Authorization Code Grant**
+  and **Hybrid** flows, which can be used to integrate with `OAuth` applications
+- Adds built-in handling of `X-Forwarded` and `Forwarded` headers with
+  validation, and also to handle trusted hosts
+- Adds an extensibility point that enables sorting of middlewares before they
+  are applied on the application
 - Fixes #199
 - Downgrades `httptools` dependency to version `>=0.2,<0.4`
+- Adds some improvements to the testing `get_example_scope` method
 
 ## [1.2.0] - 2021-10-24 📦
 - Includes `Python 3.10` in the CI/CD matrix